% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rough_svg.R
\name{save_rough_html}
\alias{save_rough_html}
\title{Save a roughened plot as HTML}
\usage{
save_rough_html(plot, file, ...)
}
\arguments{
\item{plot}{A ggplot object.}

\item{file}{Output HTML file.}

\item{...}{Passed to rough_plot().}
}
\description{
Save a roughened plot as HTML
}
